Add files via upload
You can add to the list by typing in the "What needs to be done?" box and hitting the return key.  You can remove something from the list by using the "x" button at the right of every item (appears when you hover over it). When you check the circle on the left, it strikes out the task to show that the same is complete and the color also changes from Black to Grey.  Now the exercise is to modify the existing code so that the following are achieved: